requirements.txt
beamngpy==1.18
Keras==2.4.3
tensorflow==2.4.0
shapely==1.7.1
pandas==1.2.1
imageio==2.9.0
scikit-learn==0.24.1
opencv-python==4.5.1.48
PyYAML~=5.3.1
numpy~=1.19.5
matplotlib~=3.3.3